models_app/requirements.txt
Flask==2.0.2
numpy==1.21.3
pandas==1.3.4
Cython==0.29.24
psycopg2-binary
scikit-learn==1.0.1
requests==2.26.0
flasgger==0.9.5